+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New command `!death` to track death-counters and get some (hopefully) funny replies
 - Support https://ao.tinkeringidiot.com/ for Discord item links
 - New command `!reck` to show mob-status in The Reck
+- New command `!hi` to show status of Hollow Island, including which wave is currently running. We're deliberately not showing any ninja-looting related information.
 
 ### Fixed